+++ actioncable/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Rails 6.1.7.9 (October 15, 2024) ##
+
+*   No changes.
+
+
 ## Rails 6.1.7.8 (June 04, 2024) ##
 
 *   No changes.